Re-enabled smoke tests

+++ src/_smoke_test.js
@@ -1,6 +1,11 @@
-// Copyright (c) 2012 Titanium I.T. LLC. All rights reserved. See LICENSE.txt for details.
+// Copyright (c) 2012-2015 Titanium I.T. LLC. All rights reserved. See LICENSE.txt for details.
+
+// Runs the smoke tests against a localhost server.
+
 (function() {
 	"use strict";
+
+	var expect = require("expect.js");
 	var child_process = require("child_process");
 	var fs = require("fs");
 	var procfile = require("procfile");
@@ -9,45 +14,48 @@
 	var PORT = "5000";
 	var BASE_URL = "http://localhost:" + PORT;
 
-	var child;
+	describe("Localhost server", function() {
 
-	exports.setUp = function(done) {
-		var stdout = "";
+		var child = null;
 
-		var web = parseProcfile();
-		child = child_process.spawn(web.command, web.options, { stdio: "pipe" });
-		child.stdout.setEncoding("utf8");
-		child.stdout.on("data", function(chunk) {
-			if (stdout !== null) stdout += chunk;
-			if (stdout.trim() === "Server started") {
-				stdout = null;
+		beforeEach(function(done) {
+			var cmdLine = parseProcfile();
+			child = child_process.spawn(cmdLine.command, cmdLine.options, { stdio: [ "pipe", "pipe", process.stderr ]});
+
+			child.stdout.setEncoding("utf8");
+			child.stdout.on("data", function(chunk) {
+				process.stdout.write(chunk);
+				if (chunk.trim().indexOf("Server started") !== -1) return done();
+			});
+
+			child.once("exit", function() {
+				child = null;
+			});
+		});
+
+		afterEach(function(done) {
+			if (child === null) return done();
+
+			child.once("exit", done);
+			child.kill();
+		});
+
+		it("passes smoke tests", function(done) {
+			smoketest.runTests(BASE_URL, function(success) {
+				expect(success).to.be(true);
 				done();
-			}
+			});
 		});
-	};
 
-	function parseProcfile() {
-		var file = fs.readFileSync("Procfile", "utf8");
-		var web = procfile.parse(file).web;
-		web.options = web.options.map(function(option) {
-			if (option === "$PORT") return PORT;
-			else return option;
-		});
-		return web;
-	}
-
-	exports.tearDown = function(done) {
-		child.on("exit", function() {
-			done();
-		});
-		child.kill();
-	};
-
-	exports.test_localhost = function(test) {
-		smoketest.runTests(BASE_URL, function(success) {
-			if (!success) test.fail("smoke tests failed");
-			test.done();
-		});
-	};
+		function parseProcfile() {
+			var file = fs.readFileSync("Procfile", "utf8");
+			var web = procfile.parse(file).web;
+			web.options = web.options.map(function(option) {
+				if (option === "$PORT") return PORT;
+				else return option;
+			});
+			return web;
+		}
+	});
 
 }());
